Show the cluster activity when the screen is locked.
In the AndroidManifest file, set `showWhenLocked` and `turnScreenOn` to
true, in order to show the cluster activity even when the screen is
locked. The behavior can be controlled by the OEMs who write the cluster
app, and we are updating the sample app as an example.
Also updating the README file with the directions.

ClusterHomeActivityLightMode is also updated to make it clear
that it is running in the LIGHT mode, and to show the heartbeats.

Bug: 335733292
Flag: EXEMPT bugfix
Test: manual experiments with Seahawk.

Change-Id: I7aea4a175a97a47371f5c338e5b52d3a6d10bd7e